requirements.txt
Django==4.2.*
djangorestframework
psycopg2-binary
Pillow
gunicorn
django-cors-headers
django-simpleui